refactor: legacy interface Controller 삭제
