+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,0 +1,25 @@
+package com.techcourse.controller;
+
+import com.techcourse.domain.User;
+import com.techcourse.repository.InMemoryUserRepository;
+import context.org.springframework.stereotype.Controller;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import web.org.springframework.web.bind.annotation.RequestMapping;
+import web.org.springframework.web.bind.annotation.RequestMethod;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.JspView;
+
+@Controller
+public class RegisterController {
+
+    @RequestMapping(value = "/register", method = {RequestMethod.POST})
+    public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) {
+        final var user = new User(2,
+            req.getParameter("account"),
+            req.getParameter("password"),
+            req.getParameter("email"));
+        InMemoryUserRepository.save(user);
+        return new ModelAndView(new JspView("/index.jsp"));
+    }
+}
