+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,7 +1,5 @@
-package com.techcourse.controller.annotaion;
+package com.techcourse.controller;
 
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,15 +9,10 @@ import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 @Controller
-public class RegisterController {
+public class RegisterViewController {
 
-    @RequestMapping(value = "/register", method = {RequestMethod.POST})
+    @RequestMapping(value = "/register/view", method = {RequestMethod.GET})
     public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) {
-        final var user = new User(2,
-            req.getParameter("account"),
-            req.getParameter("password"),
-            req.getParameter("email"));
-        InMemoryUserRepository.save(user);
-        return new ModelAndView(new JspView("/index.jsp"));
+        return new ModelAndView(new JspView("/register.jsp"));
     }
 }
