--- app/src/main/java/com/techcourse/controller/annotaion/LogoutController.java
@@ -1,23 +1,0 @@
-package com.techcourse.controller.annotaion;
-
-import com.techcourse.controller.UserSession;
-import context.org.springframework.stereotype.Controller;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import web.org.springframework.web.bind.annotation.RequestMapping;
-import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-@Controller
-public class LogoutController {
-
-
-    @RequestMapping(value = "/logout", method = {RequestMethod.GET})
-    public ModelAndView logout(HttpServletRequest req, HttpServletResponse res) {
-        final var session = req.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return new ModelAndView(new JspView("redirect:/"));
-    }
-
-}
