+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/AnnotationControllerAdapter.java
@@ -1,26 +1,24 @@
-package webmvc.org.springframework.web.servlet.mvc.asis;
+package webmvc.org.springframework.web.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspView;
 
-public class InterfaceControllerAdapter implements HandlerAdapter {
+public class AnnotationControllerAdapter implements HandlerAdapter {
 
     @Override
     public boolean support(Object object) {
-        return object instanceof Controller;
+        return object instanceof HandlerExecution;
     }
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         checkValidAdapter(handler);
-        String responsePath = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(responsePath));
+        return ((HandlerExecution) handler).handle(request, response);
     }
 
     private void checkValidAdapter(Object handler) {
-        if (!(handler instanceof Controller)) {
+        if (!(handler instanceof HandlerExecution)) {
             throw new IllegalArgumentException("해당 Adapter 는 전달된 handler 를 처리할 수 없습니다.");
         }
     }
