+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/exception/HandlerNotFoundException.java
@@ -1,0 +1,5 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe.exception;
+
+public class HandlerNotFoundException extends RuntimeException {
+
+}
